refactor: create abstract ability
