+++ config/Constants.py
@@ -105,6 +105,7 @@ class Constants:
     ABILITY_SPEED = 500
     MISSILE_SHOT_CAPACITY = 5
     ANGLE_SPREAD_MISSILE = 10
+    MISSILE_LIFETIME = 10 * WIDTH/ABILITY_SPEED
     EXPLOSION_RADIUS = 100
     COLOR_EXPLOSION = (255, 165, 0, 180)
     LASER_LIFETIME = 1 / 60
@@ -118,3 +119,4 @@ class Constants:
     COLOR_GLOW_CRITICAL_SHOT = (255, 255, 180, 160)
     CRITICAL_SHOT_WIDTH_BORDER = 24
     CRITICAL_SHOT_HEIGHT_BORDER = 24
+    CRITICAL_SHOT_LIFETIME = WIDTH/ABILITY_SPEED
